# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
      "ChecksumResponse": {
        "description": "Checksum metadata attached to stored content or versions.",
        "properties": {
          "algorithm": {
            "description": "Checksum algorithm used by the service.",
            "title": "Algorithm",
            "type": "string"
          },
          "digest": {
            "description": "Hex digest returned by the service.",
            "title": "Digest",
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "digest"
        ],
        "title": "ChecksumResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProvenanceResponse": {
        "description": "Minimal provenance returned by immutable version reads.",
        "properties": {
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          },
          "tree_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tree Path"
          }
        },
        "required": [
          "repo_url",
          "commit_sha"
        ],
        "title": "ProvenanceResponse",
        "type": "object"
      },
...
      "SkillContentSummaryResponse": {
        "description": "Compact content metadata returned without the full markdown body.",
        "properties": {
          "checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          },
          "rendered_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional pre-rendered short summary stored with the markdown.",
            "title": "Rendered Summary"
          },
          "size_bytes": {
            "description": "UTF-8 byte length of the stored markdown.",
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "checksum",
          "size_bytes",
          "rendered_summary"
        ],
        "title": "SkillContentSummaryResponse",
        "type": "object"
      },
...
      "SkillDiscoveryRequest": {
        "additionalProperties": false,
        "description": "Body-based discovery request.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "SkillDiscoveryRequest",
        "type": "object"
      },
      "SkillDiscoveryResponse": {
        "description": "Ordered candidate slugs returned by discovery.",
        "properties": {
          "candidates": {
            "items": {
              "type": "string"
            },
            "title": "Candidates",
            "type": "array"
          }
        },
        "required": [
          "candidates"
        ],
        "title": "SkillDiscoveryResponse",
        "type": "object"
      },
...
      "SkillMetadataResponse": {
        "description": "Full normalized metadata block returned by immutable metadata reads.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "inputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputs Schema"
          },
          "maturity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maturity Score"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Schema"
          },
          "security_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Score"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "token_estimate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Estimate"
          }
        },
        "required": [
          "name",
          "description",
          "tags"
        ],
        "title": "SkillMetadataResponse",
        "type": "object"
      },
      "SkillVersionBatchRequest": {
        "additionalProperties": false,
        "description": "Ordered exact immutable coordinate batch request.",
        "properties": {
          "coordinates": {
            "items": {
              "$ref": "#/components/schemas/SkillVersionCoordinateRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Coordinates",
            "type": "array"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "SkillVersionBatchRequest",
        "type": "object"
      },
...
      "SkillVersionCoordinateRequest": {
        "additionalProperties": false,
        "description": "Exact immutable slug/version coordinate.",
        "properties": {
          "slug": {
            "description": "Stable public slug of the requested skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "description": "Exact immutable semantic version of the requested skill.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version"
        ],
        "title": "SkillVersionCoordinateRequest",
        "type": "object"
      },
...
      "SkillVersionMetadataBatchItemResponse": {
        "description": "One ordered immutable metadata batch result.",
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/SkillVersionCoordinateRequest"
          },
          "item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkillVersionMetadataResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "found",
              "not_found"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "coordinate"
        ],
        "title": "SkillVersionMetadataBatchItemResponse",
        "type": "object"
      },
      "SkillVersionMetadataBatchResponse": {
        "description": "Ordered immutable metadata batch response.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SkillVersionMetadataBatchItemResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SkillVersionMetadataBatchResponse",
        "type": "object"
      },
...
      "SkillVersionMetadataResponse": {
        "description": "Immutable metadata envelope returned by publish and metadata batch fetch.",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SkillContentSummaryResponse"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/SkillMetadataResponse"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "version_checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          }
        },
        "required": [
          "slug",
          "version",
          "version_checksum",
          "content",
          "metadata",
          "lifecycle_status",
          "trust_tier",
          "published_at"
        ],
        "title": "SkillVersionMetadataResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, candidate discovery,\nexact dependency reads, immutable metadata batch fetch, immutable content batch\nfetch, and governed lifecycle updates.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/discovery": {
      "post": {
        "description": "Return ordered candidate slugs from indexed metadata and description search. This route is discovery-only: it does not resolve dependencies, choose a final candidate, or perform solver behavior.",
        "operationId": "discoverSkillCandidates",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "description": "Lint Python files consistently",
                "name": "Python Lint",
                "tags": [
                  "python",
                  "lint"
                ]
              },
              "schema": {
                "$ref": "#/components/schemas/SkillDiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "candidates": [
                    "python.lint",
                    "python.format"
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillDiscoveryResponse"
                }
              }
            },
            "description": "Ordered candidate slugs returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The discovery request is invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Discover candidate skill slugs",
        "tags": [
          "discovery"
        ]
      }
    },
    "/fetch/content:batch": {
      "post": {
        "description": "Return immutable markdown content as `multipart/mixed` in request order. Each part includes coordinate and status headers.",
        "operationId": "batchGetImmutableContent",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "coordinates": [
                  {
                    "slug": "python.lint",
                    "version": "1.2.3"
                  },
                  {
                    "slug": "python.missing",
                    "version": "9.9.9"
                  }
                ]
              },
              "schema": {
                "$ref": "#/components/schemas/SkillVersionBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "multipart/mixed": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Ordered immutable markdown parts returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body is invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch immutable markdown content in batch",
        "tags": [
          "fetch"
        ]
      }
    },
    "/fetch/metadata:batch": {
      "post": {
        "description": "Return immutable metadata envelopes for an ordered list of exact coordinates. Missing coordinates are returned as `not_found` items.",
        "operationId": "batchGetImmutableMetadata",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "coordinates": [
                  {
                    "slug": "python.lint",
                    "version": "1.2.3"
                  },
                  {
                    "slug": "python.missing",
                    "version": "9.9.9"
                  }
                ]
              },
              "schema": {
                "$ref": "#/components/schemas/SkillVersionBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "results": [
                    {
                      "coordinate": {
                        "slug": "python.lint",
                        "version": "1.2.3"
                      },
                      "item": {
                        "content": {
                          "checksum": {
                            "algorithm": "sha256",
                            "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                          },
                          "rendered_summary": "Lint Python files consistently.",
                          "size_bytes": 123
                        },
                        "lifecycle_status": "published",
                        "metadata": {
                          "description": "Linting skill",
                          "headers": {
                            "runtime": "python"
                          },
                          "inputs_schema": {
                            "type": "object"
                          },
                          "maturity_score": 0.9,
                          "name": "Python Lint",
                          "outputs_schema": {
                            "type": "object"
                          },
                          "security_score": 0.95,
                          "tags": [
                            "python",
                            "lint"
                          ],
                          "token_estimate": 128
                        },
                        "provenance": {
                          "commit_sha": "aabbccddeeff00112233445566778899aabbccdd",
                          "repo_url": "https://github.com/example/skills",
                          "tree_path": "skills/python.lint"
                        },
                        "published_at": "2026-03-10T08:30:00Z",
                        "slug": "python.lint",
                        "trust_tier": "internal",
                        "version": "1.2.3",
                        "version_checksum": {
                          "algorithm": "sha256",
                          "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                        }
                      },
                      "status": "found"
                    },
                    {
                      "coordinate": {
                        "slug": "python.missing",
                        "version": "9.9.9"
                      },
                      "status": "not_found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionMetadataBatchResponse"
                }
              }
            },
            "description": "Ordered immutable metadata results returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body is invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch immutable metadata in batch",
        "tags": [
          "fetch"
        ]
      }
    },
...
  }
}